Add function to track connections and retrieve users in a document
This change adds functions to track connections to a document and retrieve the users in that document. The `adicionarConexao` function is used to add a connection to the list, while the `obterUsuariosDocumento` function is used to retrieve the users in a document by filtering the connections based on the document name and mapping to the usernames.